+++ lib/generators/ahoy/stores/templates/active_record_event_model.rb
@@ -3,7 +3,7 @@ module Ahoy
     self.table_name = "ahoy_events"
 
     belongs_to :visit
-    belongs_to :user<% if options["database"].blank? %>
+    belongs_to :user<% unless %w(postgresql postgresql-jsonb).include?(@database) %>
 
     serialize :properties, JSON<% end %>
   end