+++ lib/generators/ahoy/stores/templates/active_record_events_migration.rb
@@ -1,20 +1,19 @@
 class <%= migration_class_name %> < ActiveRecord::Migration
   def change
-    create_table :ahoy_events, id: false do |t|
-      t.uuid :id, default: nil, primary_key: true
-      t.uuid :visit_id, default: nil
+    create_table :ahoy_events do |t|
+      t.integer :visit_id
 
       # user
       t.integer :user_id
       # add t.string :user_type if polymorphic
 
       t.string :name
-      t.<% case options["database"] when "postgresql" %>json<% when "postgresql-jsonb" %>jsonb<% else %>text<% end %> :properties
+      t.<% case @database when "postgresql" %>json<% when "postgresql-jsonb" %>jsonb<% else %>text<% end %> :properties
       t.timestamp :time
     end
 
-    add_index :ahoy_events, [:visit_id]
-    add_index :ahoy_events, [:user_id]
-    add_index :ahoy_events, [:time]
+    add_index :ahoy_events, [:visit_id, :name]
+    add_index :ahoy_events, [:user_id, :name]
+    add_index :ahoy_events, [:name, :time]
   end
 end