+++ lib/generators/ahoy/stores/templates/active_record_initializer.rb
@@ -1,3 +1,3 @@
-class Ahoy::Store < Ahoy::Stores::ActiveRecordStore
+class Ahoy::Store < Ahoy::Stores::ActiveRecordTokenStore
   # customize here
 end